- Flujo de venta completo con lógica. - Vista final de ticket completo. - Objetos detalleTicket y TicketCompra guardan todos los datos modelados. - DetalleTicketData ahora envía un double para subtotal - TicketCompra tiene un total + I.V.A - Leves ajustes visuales

+++ GP15_c2_proyectoFinal/src/persistencia/DetalleTicketData.java
@@ -36,7 +36,7 @@ private Connection con;
             PreparedStatement psDetalle = con.prepareStatement(sqlDetalles, Statement.RETURN_GENERATED_KEYS);            
             psDetalle.setInt(1, codD);
             psDetalle.setInt(2, dt.getCantidad());
-            psDetalle.setInt(3, dt.getSubtotal());
+            psDetalle.setDouble(3, dt.getSubtotal());
             psDetalle.setInt(4, dt.getProyeccion().getIdProyeccion());    
             psDetalle.setDate(5, java.sql.Date.valueOf(dt.getFechProyeccion()));
